This commit adds conditions to the toggle activity functions.
This was done to ensure activity is only toggled in the event that nodes/edges exist and are currently the proper status which allows them to be toggled. (An active node cannot be marked active, etc.) This was done so we can include representative error messages in the console when the functions cannot be completed.