This commit includes cumulative weight in the shortest path return.
This allows us to easily determine the cost of the shortest path.

Furthermore, a more descript error message has been added in the event that no path exists which connects the two nodes (disconnected nodes).

Finally, the docstring associated with this function has been updated to reflect the changes in this commit.