Fixes condition for disconnected nodes
Previous condition was never satisfied which resulted in invalid paths being returned. This patch fixes that and alters the 'disconnected' return to match the structure of the conncted return (path, weight --> None, inf). This also matches the structure of the static dijkstra function. This allows for consistent unpacking.

Furthermore, this fix also simplifies the initializer.